Rename file not.hpp -> logical_not.hpp
