+++ src/core/reference/include/openvino/reference/logical_not.hpp
